数据库查询优化 - 其实是算法优化

+++ src/main/java/com/vefuture/big_bottle/web/vefuture/entity/vo/B3tyTokenTransDto.java
@@ -2,14 +2,16 @@ package com.vefuture.big_bottle.web.vefuture.entity.vo;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 /**
  * @author wangb
  * @date 2025/5/19
- * @description TODO: 类描述
+ * @description 发币时用的对象
  */
 @Data
 public class B3tyTokenTransDto {
     private String  walletAddress;
-    private Integer b3tyToken;
+    private BigDecimal b3tyToken;
     private String imgUrl;
 }
