+++ src/main/java/com/vefuture/big_bottle/web/vefuture/entity/vo/StatisticsResultDTO.java
@@ -30,5 +30,5 @@ public class StatisticsResultDTO {
     //系数
     private BigDecimal conversionFactor;
     //本次计算出的总积分
-    private Integer sumToken;
+    private BigDecimal sumToken;
 }
